Use `u16` for progress and quality to reduce memory usage (#323)

+++ raphael-sim/tests/action_tests.rs
@@ -19,7 +19,7 @@ const SETTINGS: Settings = Settings {
 /// - Quality
 /// - Durability (used)
 /// - CP (used)
-fn primary_stats(state: &SimulationState, settings: &Settings) -> (u32, u32, u16, u16) {
+fn primary_stats(state: &SimulationState, settings: &Settings) -> (u16, u16, u16, u16) {
     (
         state.progress,
         state.quality,
@@ -500,7 +500,7 @@ fn test_trained_eye() {
         .unwrap();
     assert_eq!(
         primary_stats(&state, &SETTINGS),
-        (0, u32::from(SETTINGS.max_quality), 10, 250)
+        (0, SETTINGS.max_quality, 10, 250)
     );
     assert_eq!(state.effects.inner_quiet(), 1);
 }
